Create MainInterFace
Suppose you are a software developer trying to launch your own online social network (OSN) 
(similar to Facebook or LinkedIn). The OSN should provide options for users or organizations to 
join by creating profiles/pages. Users can become friends with each other, or follow specific 
pages. All users and organizations can post text based messages which are displayed on the 
profile page, and friends/followers should be able to get a feed from their connections. 
a) Given the above scenario, identify and name the different classes that can be used to 
represent the system.
b) Provide UML diagrams for the classes identified in (a).
c) Implement a restricted version of the program that assumes only 5 users/pages can be part of
the OSN. In this version, each user can have a maximum of 3 posts/updates. The feed feature 
is disabled; so to check updates, you need to visit the profile of user/organization.